requirements.txt
phonenumbers==8.12.45
python-telegram-bot==11.1.0
